Revert "change indexnumX num[] to u32"
This reverts commit 046471c07b9b18f1542c06ca33367c607b5fee63.

+++ 2048.cpp
@@ -461,7 +461,7 @@ u64 indexmerge1(const board& b) { // 8-bit
 
 u64 indexnum0(const board& b) { // 10-bit
 	// 2k ~ 32k, 2-bit ea.
-	u32 num[16];
+	u16 num[16];
 	b.numof(num, 11, 16);
 	register u64 index = 0;
 	index += (num[11] & 0x03) << 0;
@@ -473,7 +473,7 @@ u64 indexnum0(const board& b) { // 10-bit
 }
 
 u64 indexnum1(const board& b) { // 25-bit
-	u32 num[16];
+	u16 num[16];
 	b.numof(num, 5, 16);
 	register u64 index = 0;
 	index += ((num[5] + num[6]) & 0x0f) << 0; // 32 & 64, 4-bit
@@ -490,7 +490,7 @@ u64 indexnum1(const board& b) { // 25-bit
 }
 
 u64 indexnum2(const board& b) { // 25-bit
-	u32 num[16];
+	u16 num[16];
 	b.numof(num, 0, 16);
 	register u64 index = 0;
 	index += ((num[1] + num[2]) & 0x07) << 0; // 2 & 4, 3-bit
@@ -530,7 +530,7 @@ u64 indexnum2x(const board& b) { // 25-bit
 }
 
 u64 indexnum3(const board& b) { // 28-bit
-	u32 num[16];
+	u16 num[16];
 	b.numof(num, 0, 16);
 	register u64 index = 0;
 	index += ((num[0] + num[1] + num[2]) & 0x0f) << 0; // 0 & 2 & 4, 4-bit
